+++ pkgs/code_assets/lib/code_assets.dart
@@ -103,11 +103,6 @@ export 'src/code_assets/link_mode.dart'
         LookupInProcess,
         StaticLinking;
 export 'src/code_assets/link_mode_preference.dart' show LinkModePreference;
-export 'src/code_assets/native_library_validation.dart'
-    show
-        NativeLibraryValidation,
-        NativeLibraryValidationContext,
-        NativeLibraryValidator;
 export 'src/code_assets/os.dart' show OS;
 export 'src/code_assets/sanitizer.dart' show Sanitizer;
 export 'src/code_assets/testing.dart' show testCodeBuildHook;