+++ pkgs/hooks_runner/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Revalidate cached hook output and invalidate output and dependency hashes
   when it can no longer be read or fails current validation.
 - Bump the dependency on `package:code_assets` to `^1.3.0-wip`.
+- Bump the dependency on `package:hooks` to `^2.2.0-wip`.
 - Require `dartExecutable` to be an absolute path, and on Windows to include a
   file extension